feat: update queries, add health check, fix handlers
* Update SQL queries, change Query with Exec on inserts
* Fix nil types on handlers
* Add /profile/:guid endpoint

+++ pkg/leaderboard/types.go
@@ -1,7 +1,7 @@
 package leaderboard
 
 type ScoreSubmit struct {
-	ScoreWorth float32 `json:"score_worth"`
+	ScoreWorth float64 `json:"score_worth"`
 	UserID     string  `json:"user_id"`
 }
